feat: link dotfiles to local repo on install
Instead of copying over files from a local checkout to the user
home directory, create symlinks or link directly to the dotfiles
directory checkout.

This will make it easier to pull in changes to dotfiles, as the
install script will not need to be run again as the changes
should get picked up the next time the user logs in or sources
their bash files.

Addresses #6